+++ CHANGELOG.md
@@ -11,4 +11,4 @@
 # [1.0.0] - 2025-02-16 
 - Formulario modal para enviar dados de adicionar produto
 # [1.0.0] - 2025-02-18
-- Teste de formulario
+- Modificação do modal/formulario